Refactoring client code.
Making test_client its own module, and give people the ability to
specify how many iterations to do (default to 1000).

Now client module is a base class which encapsulates the logic for
emitting to Firetower in the correct format.

Changed the entry point for firetower-client to point to the test_client
code.